1.修正部分编结警告 2.去除utilites工程文件中无用的stdafx.h头文件引入 3.完全适配gn构建。

+++ trunk/SOUI/src/control/SComboBase.cpp
@@ -252,7 +252,7 @@ namespace SOUI
         if (!m_bDropdown && m_pEdit)
         {
             m_pEdit->SetFocus();
-            m_pEdit->SetSel(MAKELONG(0,-1));
+            m_pEdit->SetSel((DWORD)MAKELONG(0,-1));
         }
 
         m_dwBtnState = WndState_Normal;
